Remove cancelOrderNoThrow tests

+++ contracts/exchange/contracts/src/MixinExchangeCore.sol
@@ -257,7 +257,7 @@ contract MixinExchangeCore is
         _assertValidCancel(order, orderInfo);
 
         // Noop if order is already unfillable
-        if (orderInfo.orderStatus != OrderStatus.FILLABLE) {
+        if (orderInfo.orderStatus != uint8(OrderStatus.FILLABLE)) {
             return;
         }
